requirements.txt
flask
python-dotenv
flask-cors
requests
pandas
snowflake-connector-python